fix(ci): lint the GUI, which nothing was linting
clang-tidy failed on the terrain tests over bugprone-unchecked-optional-
access. The check does not follow ASSERT_TRUE - a gtest macro hides the
relationship between the assertion and has_value() - and value() is no better,
because throwing counts as unchecked access too. A plain `if (!x.has_value())
{ FAIL(); }` is understood, and now reads as an assertion to more than a
person.

Running that check locally rather than pushing and waiting turned up the
larger problem. The ci-tidy preset builds with the GUI off, so its compile
database holds one file under app/ - the command line entry point - and none
of viz/src/MapScene.cpp. Every GUI and map-scene source has been outside the
lint gate since the map view was added, while the job's file pattern named
app and viz as though they were covered.

Turning it on found five errors: a vtkIdType narrowed to int, two parameters
copied for every call and only read, a function-local static named as a
constant, and a push_back loop with no reserve. All fixed here.

The gate now runs in the map-view job, where Qt and VTK are already installed
and installing them in the tidy job would cost more than this whole job takes.
That job builds with clang for the same reason ci-tidy does: the compile
database has to match the clang-tidy that reads it.

One finding is left alone and belongs to neither entry point's author: both
mains catch std::exception, so a foreign exception escapes and
bugprone-exception-escape says so. It predates this work, it is identical in
both, and changing what an entry point does on an unknown throw is its own
decision.

371 tests pass, format, line endings, dependency direction and clang-tidy all
clean locally.

+++ viz/src/MapScene.cpp
@@ -57,7 +57,10 @@ void build_rings(const std::vector<core::Polygon>& boundaries,
       points->InsertNextPoint(vertex.easting, vertex.northing, 0.1);
     }
     const auto count = static_cast<vtkIdType>(vertices.size());
-    lines->InsertNextCell(count + 1);
+    // InsertNextCell counts points in an int while the ids are vtkIdType, which
+    // is 64-bit. A paddock has a handful of corners, so the cast is safe here
+    // and saying so explicitly is better than letting it narrow silently.
+    lines->InsertNextCell(static_cast<int>(count + 1));
     for (vtkIdType i = 0; i < count; ++i) {
       lines->InsertCellPoint(first + i);
     }
